userregistration/requirements.txt
django>=1.9.0
django-rest-auth==0.9.3
djangorestframework>=3.7.0
django-allauth>=0.24.1